Move httpclient to auto configuration bean
